feat(dogstatsd): add shared runtime sink
