+++ libdd-dogstatsd-client/Cargo.toml
@@ -18,11 +18,15 @@ serde = { version = "1.0", features = ["derive", "rc"] }
 tracing = { version = "0.1", default-features = false }
 anyhow = { version = "1.0" }
 http = "1.1"
+libdd-shared-runtime = { version = "1.0.0", path = "../libdd-shared-runtime", optional = true }
+tokio = { version = "1.23", features = ["sync"], optional = true }
+async-trait = { version = "0.1", optional = true }
 
 [features]
-default = ["https"]
+default = ["https", "shared-runtime"]
 https = ["libdd-common/https"]
 fips = ["libdd-common/fips"]
+shared-runtime = ["dep:libdd-shared-runtime", "dep:tokio", "dep:async-trait"]
 
 
 [dev-dependencies]
